requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
sqlparse==0.5.1
tzdata==2024.2
django-import-export
